Created fragments
I created all the front-end fragments I think I need.  I created a
fragment for the login screen, the screen that displays all the
mistakes, the screen that is used to edit/delete/select an existing
fragment (the edit one), and a screen to insert mistakes.  I also did
some ghetto state management by having my program only load fragments if
the statemanagement thing is null.  I don't know if that's the correct
or best way to do it but I'm not getting errors when I rotate my screen
so I'm leaving it for now and I'll come back if I have problems.  I
still haven't created any Java classes for the back end which I'll need
to do.  Specifically I'll need some classes to manage the database.