Updated navigation
I updated the navigation of the app.  Now when users long-press or
press an item on the menu it loads the edit mistake page.  In addition
I changed the name of the method that loads the mistake list to
openMistakeList() instead of successfulLogin since the new name is
easier to understand.